create a central lambda wrapper handler
