+++ lambda/create.sh
@@ -3,7 +3,8 @@
 ##
 # Create a Lambda function from a zip file.
 
-FUNCTION_NAME=$1
+SCANNER_NAME=$1
+FUNCTION_NAME="task_$SCANNER_NAME"
 
 echo "Creating function: $FUNCTION_NAME"
 
@@ -13,7 +14,7 @@ aws lambda create-function \
   --function-name $FUNCTION_NAME \
   --zip-file fileb://./$FUNCTION_NAME.zip \
   --role $AWS_LAMBDA_ROLE \
-  --handler $FUNCTION_NAME.handler \
+  --handler lambda_handler.handler \
   --runtime python3.6 \
   --timeout 30 \
   --memory-size 128
